requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
Pillow>=10.0.0
python-pptx>=0.6.21
scikit-learn>=1.3.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1
scipy>=1.11.0
statsmodels>=0.14.0
groq>=0.3.0
tqdm>=4.66.0
joblib>=1.3.0
colorama>=0.4.6